Started cleaning up codebase
This is just a small cleanup, bigger ones will follow
